+++ includes/class-wizards.php
@@ -30,9 +30,6 @@ class Wizards {
 	public static function init() {
 		self::$wizards = [
 			'components-demo'         => new Components_Demo(),
-			'seo'                     => new SEO_Wizard(),
-			'health-check'            => new Health_Check_Wizard(),
-			'engagement'              => new Engagement_Wizard(),
 			// v2 Information Architecture.
 			'newspack-dashboard'      => new Newspack_Dashboard(),
 			'setup'                   => new Setup_Wizard(),
@@ -43,6 +40,7 @@ class Wizards {
 						'social-pixels' => 'Newspack\Wizards\Newspack\Pixels_Section',
 						'recirculation' => 'Newspack\Wizards\Newspack\Recirculation_Section',
 						'syndication'   => 'Newspack\Wizards\Newspack\Syndication_Section',
+						'seo'           => 'Newspack\Wizards\Newspack\Seo_Section',
 					],
 				]
 			),