+++ includes/cli/class-woocommerce-subscriptions.php
@@ -69,6 +69,7 @@ class WooCommerce_Subscriptions {
 		self::$verbose = isset( $assoc_args['verbose'] ) ? true : false;
 		$scheduled     = 0;
 		$updated       = 0;
+		$trashed       = 0;
 		$page          = 1;
 		$per_page      = 25;
 		$subscriptions = self::get_subscriptions( $page );
@@ -93,7 +94,7 @@ class WooCommerce_Subscriptions {
 					}
 					continue;
 				}
-				$renewal_order = $subscription->get_last_order(
+				$last_order = $subscription->get_last_order(
 					'all',
 					[ 'renewal' ],
 					[
@@ -102,85 +103,82 @@ class WooCommerce_Subscriptions {
 						'refunded',
 					]
 				);
-				// No failed or pending renewal orders indicates the subscription was likely manually placed on hold.
-				if ( empty( $renewal_order ) ) {
-					if ( self::$verbose ) {
-						WP_CLI::line( 'Subscription has no pending renewal orders. Moving to next subscription...' );
-						WP_CLI::line( '' );
+				if ( ! $last_order ) {
+					$last_order = $subscription->get_parent();
+					// If the last order is the parent order and has a failed status, trash the subscription.
+					if ( $last_order && 'failed' === $last_order->get_status() ) {
+						if ( self::$verbose ) {
+							WP_CLI::line( 'Subscription parent order failed. Moving to trash...' );
+							WP_CLI::line( '' );
+						}
+						if ( self::$live ) {
+							$subscription->update_status( 'trash', __( 'Subscription status updated by Newspack CLI command.', 'newspack-plugin' ) );
+							$subscription->set_end_date( $subscription->get_date( 'next_payment' ) );
+							$subscription->update_meta_data( '_newspack_cli_status_updated', true );
+							$subscription->save();
+						}
+						++$trashed;
+						continue;
 					}
-					continue;
 				}
-				if ( ! $subscription->payment_method_supports( 'subscription_date_changes' ) ) {
-					if ( self::$verbose ) {
-						WP_CLI::line( 'Subscription payment method does not support retries. Moving to next subscription...' );
-						WP_CLI::line( '' );
-					}
-					continue;
-				}
-
 				if ( $subscription->is_manual() ) {
 					$end_date = $subscription->get_date( 'next_payment' );
 					$should_expire = wcs_date_to_time( $end_date ) + ( On_Hold_Duration::get_on_hold_duration() * DAY_IN_SECONDS ) < time();
-
+					// If the manual subscription is within the on-hold duration, schedule expiration.
 					if ( ! $should_expire ) {
 						if ( self::$verbose ) {
-							WP_CLI::line( 'Subscription is within the on-hold duration. Moving to next subscription...' );
-							WP_CLI::line( '' );
-						}
-						continue;
-					}
-				} else {
-					$last_retry = \WCS_Retry_Manager::store()->get_last_retry_for_order( wcs_get_objects_property( $renewal_order, 'id' ) );
-
-					// No retries indicates the subscription was likely manually placed on hold.
-					if ( empty( $last_retry ) ) {
-						if ( self::$verbose ) {
-							WP_CLI::line( 'No retries scheduled. Moving to next subscription...' );
-							WP_CLI::line( '' );
-						}
-						continue;
-					}
-
-					$end_date = $last_retry->get_date();
-					$should_expire = wcs_date_to_time( $end_date ) + ( On_Hold_Duration::get_on_hold_duration() * DAY_IN_SECONDS ) < time();
-
-					// A non failed status indicates the retry was either manually cancelled
-					// or was successful at one point but likely placed on hold for some other reason.
-					if ( 'failed' !== $last_retry->get_status() ) {
-						if ( self::$verbose ) {
-							WP_CLI::line( 'Last retry does not have a failed status. Moving to next subscription...' );
-							WP_CLI::line( '' );
-						}
-						continue;
-					}
-					if ( ! $should_expire ) {
-						if ( self::$verbose ) {
-							WP_CLI::line( 'Retry date is within the on-hold duration. Scheduling final retry...' );
+							WP_CLI::line( 'Manual subscription is within the on-hold duration. Scheduling expiration...' );
 						}
 						if ( self::$live ) {
-							// Retry rules can only be applied when payment attempt flag is set.
-							add_filter( 'wcs_is_scheduled_payment_attempt', '__return_true' );
-							\WCS_Retry_Manager::maybe_apply_retry_rule( $subscription, $renewal_order );
-							remove_filter( 'wcs_is_scheduled_payment_attempt', '__return_true' );
-							if ( 0 === $subscription->get_date( 'payment_retry' ) ) {
-								if ( self::$verbose ) {
-									WP_CLI::error( 'Failed to schedule payment retry. Moving to next subscription...' );
-									WP_CLI::line( '' );
-								}
-								continue;
-							} else {
-								$subscription->add_order_note(
-									__( 'Final payment retry scheduled by Newspack CLI command.', 'newspack-plugin' )
-								);
-								$subscription->update_meta_data( '_newspack_cli_retry_scheduled', true );
-								$subscription->save();
-							}
+							On_Hold_Duration::maybe_schedule_expiration( $subscription );
 						}
 						++$scheduled;
 					}
+				} else {
+					$last_retry    = \WCS_Retry_Manager::store()->get_last_retry_for_order( wcs_get_objects_property( $last_order, 'id' ) );
+					$end_date      = $last_retry ? $last_retry->get_date() : $subscription->get_date( 'next_payment' );
+					$should_expire = wcs_date_to_time( $end_date ) + ( On_Hold_Duration::get_on_hold_duration() * DAY_IN_SECONDS ) < time();
+					if ( ! $should_expire ) {
+						// If there have been retries, schedule the final retry.
+						if ( $last_retry ) {
+							if ( self::$verbose ) {
+								WP_CLI::line( 'Retry date is within the on-hold duration. Scheduling final retry...' );
+							}
+							if ( self::$live ) {
+								// Retry rules can only be applied when payment attempt flag is set.
+								add_filter( 'wcs_is_scheduled_payment_attempt', '__return_true' );
+								\WCS_Retry_Manager::maybe_apply_retry_rule( $subscription, $last_order );
+								remove_filter( 'wcs_is_scheduled_payment_attempt', '__return_true' );
+								if ( 0 === $subscription->get_date( 'payment_retry' ) ) {
+									if ( self::$verbose ) {
+										WP_CLI::error( 'Failed to schedule payment retry. Moving to next subscription...' );
+										WP_CLI::line( '' );
+									}
+									continue;
+								} else {
+									$subscription->add_order_note(
+										__( 'Final payment retry scheduled by Newspack CLI command.', 'newspack-plugin' )
+									);
+									$subscription->update_meta_data( '_newspack_cli_retry_scheduled', true );
+									$subscription->save();
+								}
+							}
+							++$scheduled;
+						} else {
+							// If there have been no retries, schedule expiration.
+							if ( self::$verbose ) {
+								WP_CLI::line( 'No retries found. Scheduling subscription expiration...' );
+							}
+							if ( self::$live ) {
+								$on_hold_duration = On_Hold_Duration::get_on_hold_duration() * DAY_IN_SECONDS;
+								On_Hold_Duration::schedule_expiration( $subscription->get_id(), $subscription->get_time( 'next_payment' ) + $on_hold_duration );
+							}
+							++$scheduled;
+						}
+					}
 				}
+				// Expire any subscriptinos that have passed the on-hold duration.
 				if ( $should_expire ) {
-					// Otherwise, if the retry date is past the on-hold duration, update the subscription status to expired.
 					if ( self::$verbose ) {
 						WP_CLI::line( 'Updating subscription status to expired...' );
 					}
@@ -199,7 +197,7 @@ class WooCommerce_Subscriptions {
 			}
 			$subscriptions = self::get_subscriptions( ++$page );
 		}
-		WP_CLI::success( 'Finished processing subscriptions. ' . $updated . ' subscriptions updated. ' . $scheduled . ' retries scheduled.' );
+		WP_CLI::success( 'Finished processing subscriptions. ' . $updated . ' subscriptions updated. ' . $scheduled . ' retries scheduled. ' . $trashed . ' subscriptions trashed.' );
 		if ( ! self::$live ) {
 			WP_CLI::warning( 'Dry run. Use --live flag to process live subscriptions.' );
 		}