+++ src/wizards/audience/types/index.d.ts
@@ -73,6 +73,7 @@ type ConfigKey = keyof Config;
 // Props for the Prequisite component.
 type PrequisiteProps = {
 	config: Config;
+	slug: string;
 	getSharedProps: (
 		configKey: string,
 		type: string
@@ -84,6 +85,12 @@ type PrequisiteProps = {
 	};
 	inFlight: boolean;
 	saveConfig: ( config: Config ) => void;
+	skipPrerequisite: (
+		data: {
+			prerequisite: string;
+			skip: boolean;
+		}
+	) => void;
 
 	// Schema for prequisite object is defined in PHP class Reader_Activation::get_prerequisites_status().
 	prerequisite: {
@@ -108,6 +115,7 @@ type PrequisiteProps = {
 		disabled_text?: string;
 		is_unavailable?: boolean;
 		is_skipped?: boolean;
+		skippable?: boolean;
 	};
 };