+++ src/wizards/audience/views/setup/campaign.js
@@ -16,7 +16,6 @@ import {
 	Notice,
 	Waiting,
 	withWizardScreen,
-	utils,
 } from '../../../../components/src';
 import Prompt from '../../components/prompt';
 import Router from '../../../../components/src/proxied-imports/router';
@@ -24,25 +23,18 @@ import './style.scss';
 
 const { useHistory } = Router;
 
-export default withWizardScreen( () => {
-	const { is_skipped_campaign_setup, reader_activation_url } =
-		newspackAudience;
-
+const AudienceCampaign = withWizardScreen( ( { error, setError, skipPrerequisite } ) => {
+	const { reader_activation_url } = newspackAudience;
 	const [ inFlight, setInFlight ] = useState( false );
-	const [ error, setError ] = useState( false );
 	const [ prompts, setPrompts ] = useState( null );
 	const [ allReady, setAllReady ] = useState( false );
-	const [ skipped, setSkipped ] = useState( {
-		status: '',
-		isSkipped: is_skipped_campaign_setup === '1',
-	} );
 	const history = useHistory();
 
 	const fetchPrompts = () => {
 		setError( false );
 		setInFlight( true );
 		apiFetch( {
-			path: '/newspack-popups/v1/reader-activation/campaign',
+			path: '/newspack-popups/v1/audience-management/campaign',
 		} )
 			.then( fetchedPrompts => {
 				setPrompts( fetchedPrompts );
@@ -50,48 +42,6 @@ export default withWizardScreen( () => {
 			.catch( setError )
 			.finally( () => setInFlight( false ) );
 	};
-
-	/**
-	 * Display prompt requiring editors to confirm skipping, on confirmation send request to
-	 * server to store skipped option in options table and redirect back to RAS
-	 *
-	 * @return {void}
-	 */
-	async function onSkipCampaignSetup() {
-		if (
-			! utils.confirmAction(
-				__(
-					'Are you sure you want to skip setting up a reader activation campaign?',
-					'newspack-plugin'
-				)
-			)
-		) {
-			return;
-		}
-		setError( false );
-		setSkipped( { ...skipped, status: 'pending' } );
-		try {
-			const request = await apiFetch( {
-				path: '/newspack/v1/wizard/newspack-audience/reader-activation/skip-campaign',
-				method: 'POST',
-				data: { skip: ! skipped.isSkipped },
-			} );
-			if ( ! request.updated ) {
-				setError( {
-					message: __( 'Server not updated', 'newspack-plugin' ),
-				} );
-				setSkipped( { isSkipped: false, status: '' } );
-				return;
-			}
-			setSkipped( { isSkipped: Boolean( request.skipped ), status: '' } );
-			newspackAudience.is_skipped_campaign_setup =
-				request.skipped ? '1' : '';
-			history.push( '/complete' );
-		} catch ( err ) {
-			setError( err );
-			setSkipped( { isSkipped: false, status: '' } );
-		}
-	}
 
 	useEffect( () => {
 		window.scrollTo( 0, 0 );
@@ -107,7 +57,7 @@ export default withWizardScreen( () => {
 	return (
 		<WizardsTab
 			title={ __(
-				'Set Up Reader Activation Campaign',
+				'Set Up Audience Management Campaign',
 				'newspack-plugin'
 			) }
 			description={ __(
@@ -142,26 +92,24 @@ export default withWizardScreen( () => {
 				) ) }
 			<div className="newspack-buttons-card">
 				<Button
-					isTertiary
-					disabled={
-						inFlight ||
-						skipped.isSkipped ||
-						skipped.status === 'pending'
-					}
-					onClick={ onSkipCampaignSetup }
+					variant={ 'secondary' }
+					isDestructive
+					disabled={ inFlight }
+					onClick={ () => {
+						skipPrerequisite(
+							{
+								prerequisite: 'ras_campaign',
+								skip: true,
+							},
+							() => history.push( '/complete' )
+						);
+					} }
 				>
-					{ /* eslint-disable-next-line no-nested-ternary */ }
-					{ skipped.status === 'pending'
-						? __( 'Skipping…', 'newspack-plugin' )
-						: skipped.isSkipped
-						? __( 'Skipped', 'newspack-plugin' )
-						: __( 'Skip', 'newspack-plugin' ) }
+					{ __( 'Skip', 'newspack-plugin' ) }
 				</Button>
 				<Button
 					isPrimary
-					disabled={
-						inFlight || ( ! allReady && ! skipped.isSkipped )
-					}
+					disabled={ inFlight || ! allReady }
 					href={ `${ reader_activation_url }complete` }
 				>
 					{ __( 'Continue', 'newspack-plugin' ) }
@@ -177,3 +125,5 @@ export default withWizardScreen( () => {
 		</WizardsTab>
 	);
 } );
+
+export default AudienceCampaign;